% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{kaya}
\alias{kaya}
\title{Kaya Identity}
\usage{
kaya(pop, gdp, enInt, carbInt)
}
\arguments{
\item{pop}{Population size (in millions).}

\item{gdp}{GDP per capita (in 1000$/person).}

\item{enInt}{Energy Intensity (in Gigajoule/$1000GDP).}

\item{carbInt}{Carbon Intensity (in tonnes CO2/Gigajoule).}
}
\value{
Total CO2 emmissions from human sources according to the kaya identity.
}
\description{
This function calculates the total CO2 emmissions from human sources according
to the kaya identity.
See \url{https://en.wikipedia.org/wiki/Kaya_identity} for more details.
}
\examples{
# Kaya Identity for Germany
kaya(82.4, 44, 5, 0.05)
}
